+++ src/Symfony/Component/DependencyInjection/Tests/Fixtures/php/services_dedup_lazy_ghost.php
@@ -42,19 +42,6 @@ class ProjectServiceContainer extends Container
         return $factory();
     }
 
-    protected function hydrateProxy($proxy, $instance)
-    {
-        if ($proxy === $instance) {
-            return $proxy;
-        }
-
-        if (!\in_array(\get_class($instance), [\get_class($proxy), get_parent_class($proxy)], true)) {
-            throw new LogicException(sprintf('Lazy service of type "%s" cannot be hydrated because its factory returned an unexpected instance of "%s". Try adding the "proxy" tag to the corresponding service definition with attribute "interface" set to "%1$s".', get_parent_class($proxy), get_debug_type($instance)));
-        }
-
-        return \Symfony\Component\VarExporter\Hydrator::hydrate($proxy, (array) $instance);
-    }
-
     /**
      * Gets the public 'bar' shared service.
      *
@@ -63,9 +50,7 @@ class ProjectServiceContainer extends Container
     protected function getBarService($lazyLoad = true)
     {
         if (true === $lazyLoad) {
-            return $this->services['bar'] = $this->createProxy('stdClass_5a8a5eb', function () {
-                return \stdClass_5a8a5eb::createLazyGhostObject($this->getBarService(...));
-            });
+            return $this->services['bar'] = $this->createProxy('stdClass_5a8a5eb', fn () => \stdClass_5a8a5eb::createLazyGhost($this->getBarService(...)));
         }
 
         return $lazyLoad;
@@ -79,16 +64,23 @@ class ProjectServiceContainer extends Container
     protected function getFooService($lazyLoad = true)
     {
         if (true === $lazyLoad) {
-            return $this->services['foo'] = $this->createProxy('stdClass_5a8a5eb', function () {
-                return \stdClass_5a8a5eb::createLazyGhostObject($this->getFooService(...));
-            });
+            return $this->services['foo'] = $this->createProxy('stdClass_5a8a5eb', fn () => \stdClass_5a8a5eb::createLazyGhost($this->getFooService(...)));
         }
 
         return $lazyLoad;
     }
 }
 
-class stdClass_5a8a5eb extends \stdClass implements \Symfony\Component\VarExporter\LazyGhostObjectInterface
+class stdClass_5a8a5eb extends \stdClass implements \Symfony\Component\VarExporter\LazyObjectInterface
 {
-    use \Symfony\Component\VarExporter\LazyGhostObjectTrait;
+    use \Symfony\Component\VarExporter\LazyGhostTrait;
+
+    private int $lazyObjectId;
+
+    private const LAZY_OBJECT_PROPERTY_SCOPES = [];
 }
+
+// Help opcache.preload discover always-needed symbols
+class_exists(\Symfony\Component\VarExporter\Internal\Hydrator::class);
+class_exists(\Symfony\Component\VarExporter\Internal\LazyObjectRegistry::class);
+class_exists(\Symfony\Component\VarExporter\Internal\LazyObjectState::class);