+++ src/Symfony/Component/VarExporter/Tests/Fixtures/LazyGhost/MagicClass.php
@@ -9,7 +9,7 @@
  * file that was distributed with this source code.
  */
 
-namespace Symfony\Component\VarExporter\Tests\Fixtures\LazyGhostObject;
+namespace Symfony\Component\VarExporter\Tests\Fixtures\LazyGhost;
 
 class MagicClass
 {