+++ src/Symfony/Component/VarExporter/CHANGELOG.md
@@ -4,7 +4,7 @@ CHANGELOG
 6.2
 ---
 
- * Add `LazyGhostObjectTrait`
+ * Add support for lazy ghost objects and virtual proxies
  * Add `Hydrator::hydrate()`
  * Preserve PHP references also when using `Hydrator::hydrate()` or `Instantiator::instantiate()`
  * Add support for hydrating from native (array) casts